+++ scripts/update_spdlog.sh
@@ -3,10 +3,11 @@
 VERSION_TAG=$1
 
 cd third || exit
-rm -rf spdlog
+if [ -e spdlog ];then
+  rm -rf spdlog
+fi
 git clone --depth=1 --branch "$VERSION_TAG" https://github.com/gabime/spdlog.git
-cd spdlog || exit
-mkdir build && cd build || exit
+cd spdlog && mkdir build && cd build || exit
 cmake .. && cmake --build . && cp "libspdlog.a" ".."
 cd ..
 find . -maxdepth 1 ! -name "libspdlog.a" ! -name "include" -exec rm -rf {} +
